move maintable related classes to new package
