+++ src/main/java/net/sf/jabref/gui/maintable/MainTableFormat.java
@@ -13,11 +13,12 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.gui;
+package net.sf.jabref.gui.maintable;
 
 import java.util.HashMap;
 import java.util.Vector;
 
+import net.sf.jabref.gui.*;
 import net.sf.jabref.model.entry.AuthorList;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.Globals;
